requirements.txt
fastapi
uvicorn
python-decouple
motor
pytest
httpx